Move getFileHash into own file for reusability
