+++ app/src/lib/file-system.ts
@@ -3,7 +3,6 @@ import * as Path from 'path'
 import { Disposable } from 'event-kit'
 import { Tailer } from './tailer'
 import byline from 'byline'
-import * as Crypto from 'crypto'
 import { createReadStream } from 'fs'
 import { mkdtemp } from 'fs/promises'
 
@@ -78,25 +77,5 @@ export async function readPartialFile(
       })
       .on('error', reject)
       .on('end', () => resolve(Buffer.concat(chunks, total)))
-  })
-}
-
-export async function getFileHash(
-  path: string,
-  type: 'sha1' | 'sha256'
-): Promise<string> {
-  return new Promise((resolve, reject) => {
-    const hash = Crypto.createHash(type)
-    hash.setEncoding('hex')
-    const input = createReadStream(path)
-
-    hash.on('finish', () => {
-      resolve(hash.read() as string)
-    })
-
-    input.on('error', reject)
-    hash.on('error', reject)
-
-    input.pipe(hash)
   })
 }
