+++ app/src/lib/ssh/ssh-key-passphrase.ts
@@ -1,4 +1,4 @@
-import { getFileHash } from '../file-system'
+import { getFileHash } from '../get-file-hash'
 import { TokenStore } from '../stores'
 import {
   getSSHSecretStoreKey,
